fixed form settings
