+++ src/components/account/settings/settings.tsx
@@ -1,33 +1,15 @@
 "use client";
 
 import React, { useState, useEffect, useMemo } from "react";
-import { ErrorOption, useForm } from "react-hook-form";
+import { ErrorOption, FormProvider, useForm } from "react-hook-form";
 import styles from "@src/components/account/settings/settings.module.scss";
-import { Button } from "@src/shared/ui/button";
-import { Input, PhoneInput } from "@src/shared/ui/inputs";
-import { CheckboxInput } from "@src/shared/ui/inputs";
-import { PasswordInput } from "@src/shared/ui/inputs";
 import { usePatchUsersMeMutation, usePostUsersSetPasswordMutation } from "@src/redux/api/users-api-slice";
 import { UserAccount } from "@src/redux/api/generated";
 import { setUser, useUser } from "@src/redux/slices/users-slice";
 import { useAppDispatch } from "@src/redux/hooks";
-import { phoneConversion } from "@src/shared/lib/phoneMask";
 import { createNotifyModal } from "@src/redux/slices/notify-modal-slice";
-import {
-  PATTERN_EMAIL,
-  PATTERN_FULLNAME,
-  PATTERN_PASSWORD,
-  SETTINGS_EMAIL,
-  SETTINGS_NAME,
-  SETTINGS_NEW_PASSWORD,
-  SETTINGS_NEW_PASSWORD_REPEAT,
-  SETTINGS_PHONE,
-  SETTINGS_PREVIOUS_PASSWORD,
-  SETTINGS_SURNAME,
-  VALIDATIONS_EMAIL,
-  VALIDATIONS_FULLNAME,
-  VALIDATIONS_PASSWORD,
-} from "@src/shared/constants/fields";
+import { Form } from "@src/components/account/settings/form";
+import { getUser } from "./lib/get-user";
 
 export type SettingsForm = {
   name: string;
@@ -57,30 +39,7 @@ export const Settings = () => {
   const user: UserAccount | null = useUser();
   const dispatch = useAppDispatch();
 
-  const getUser = (user: UserAccount) => {
-    const { name, surname, person_telephone, email } = user;
-
-    const phoneValue = phoneConversion(person_telephone || "");
-
-    return {
-      name: name || "",
-      surname: surname || "",
-      person_telephone: phoneValue || "",
-      email: email || "",
-      notifications: noticeEmail === null ? "off" : "email",
-    } as SettingsForm;
-  };
-
-  const {
-    register,
-    handleSubmit,
-    reset,
-    setError,
-    getValues,
-    trigger,
-    watch,
-    formState: { errors, isDirty, dirtyFields, isValid },
-  } = useForm<SettingsForm>({
+  const formMethods = useForm<SettingsForm>({
     defaultValues: useMemo(() => {
       if (user) {
         setNoticeEmail(Boolean(user?.notifications?.length) ? ["email"] : null);
@@ -90,43 +49,30 @@ export const Settings = () => {
     mode: "onChange",
   });
 
+  const {
+    handleSubmit,
+    reset,
+    setError,
+    getValues,
+    formState: { dirtyFields },
+  } = formMethods;
+
+  // TODO заменить defaultValues: useMemo на асинхронную функцию, удалить useEffect. Когда подключат кэш. Протестировать с кэшем.
   useEffect(() => {
-    if (user && !isErrorPassword) {
+    if (user && isErrorPassword) {
+      reset(
+        {
+          ...getUser(user),
+          currentPass: getValues().currentPass,
+          newPass: getValues().newPass,
+          newPassRepeat: getValues().newPassRepeat,
+        },
+        { keepErrors: true }
+      );
+    } else if (user) {
       reset(getUser(user));
     }
-  }, [user, isErrorPassword]);
-
-  // Зависимость паролей друг от друга, для дальнейшей валидации полей пароля.
-  const currentPass = watch("currentPass");
-  const newPass = watch("newPass");
-  const newPassRepeat = watch("newPassRepeat");
-  useEffect(() => {
-    if (isActiveForm) {
-      trigger("newPass");
-      trigger("currentPass");
-      trigger("newPassRepeat");
-    }
-  }, [currentPass, newPass, newPassRepeat, trigger]);
-
-  const enableForm = () => {
-    setIsActiveForm(true);
-  };
-
-  const disableForm = () => {
-    setIsActiveForm(false);
-  };
-
-  const resetForm = () => {
-    disableForm();
-    if (user) {
-      reset(getUser(user));
-    }
-    setNoticeEmail(getValues().notifications === "email" ? ["email"] : null);
-  };
-
-  const handleChangeNotice = (notice: ["email"] | null) => {
-    setNoticeEmail(notice);
-  };
+  }, [user]);
 
   const handleDefaultErrors = (data: { [x: string]: string[] }) => {
     const keys = Object.keys(data || {}) as Array<keyof SettingsForm>;
@@ -292,82 +238,11 @@ export const Settings = () => {
       }
 
       if (!user?.value?.error && !password?.value?.error) {
-        disableForm();
-        reset();
         setNoticeEmail(getValues().notifications === "email" ? ["email"] : null);
+        reset();
+        setIsActiveForm(false);
       }
     });
-  };
-
-  const buttonsFormComponent = ({
-    editedForm,
-    isLoading,
-    formParameter,
-  }: {
-    editedForm: boolean;
-    isLoading: boolean;
-    formParameter: {
-      changesForm: boolean;
-      isValidFields: boolean;
-    };
-  }) => {
-    return (
-      <div className={styles.buttonsSection}>
-        {editedForm || isLoading ? (
-          <>
-            <Button isLoading={isLoading} disabled={!formParameter.changesForm || !formParameter.isValidFields}>
-              Сохранить
-            </Button>
-
-            <Button type="button" onClick={resetForm}>
-              Отменить
-            </Button>
-          </>
-        ) : (
-          <Button type="button" onClick={enableForm}>
-            Редактировать
-          </Button>
-        )}
-      </div>
-    );
-  };
-
-  const checkboxFormComponent = ({
-    editedForm,
-    noticeEmail,
-    onClick,
-  }: {
-    editedForm: boolean;
-    noticeEmail: ["email"] | null;
-    onClick: (notice: ["email"] | null) => void;
-  }) => {
-    const onClickEmailHandler = () => onClick(["email"]);
-    const onClickOffHandler = () => onClick(null);
-
-    return (
-      <ul className={styles.list}>
-        <CheckboxInput
-          type="radio"
-          className={styles.checkbox}
-          textLabel="на e-mail"
-          disabled={!editedForm}
-          value="email"
-          {...register("notifications")}
-          checked={noticeEmail !== null}
-          onClick={onClickEmailHandler}
-        />
-        <CheckboxInput
-          type="radio"
-          className={styles.checkbox}
-          disabled={!editedForm}
-          textLabel="не получать"
-          value="off"
-          {...register("notifications")}
-          checked={noticeEmail === null}
-          onClick={onClickOffHandler}
-        />
-      </ul>
-    );
   };
 
   const deletingProfileComponent = () => {
@@ -384,169 +259,15 @@ export const Settings = () => {
   return (
     user && (
       <form onSubmit={handleSubmit(onSubmit)} className={styles.form}>
-        <div className={styles.inputsRegistration}>
-          <Input
-            label={SETTINGS_NAME.label}
-            placeholder={SETTINGS_NAME.placeholder}
-            disabled={isActiveForm === false}
-            id="firstName"
-            autoComplete="on"
-            error={Boolean(errors.name?.message)}
-            errorMessage={errors.name?.message}
-            {...register("name", {
-              required: VALIDATIONS_FULLNAME.required,
-              minLength: { ...VALIDATIONS_FULLNAME.minLength },
-              maxLength: { ...VALIDATIONS_FULLNAME.maxLength },
-              pattern: { ...PATTERN_FULLNAME },
-            })}
+        <FormProvider {...formMethods}>
+          <Form
+            noticeEmail={noticeEmail}
+            setNoticeEmail={setNoticeEmail}
+            isActiveForm={isActiveForm}
+            setIsActiveForm={setIsActiveForm}
+            isLoading={isLoadingUsers || isLoadingPassword}
           />
-
-          <Input
-            label={SETTINGS_SURNAME.label}
-            placeholder={SETTINGS_SURNAME.placeholder}
-            disabled={isActiveForm === false}
-            id="lastName"
-            autoComplete="on"
-            error={Boolean(errors.surname?.message)}
-            errorMessage={errors.surname?.message}
-            {...register("surname", {
-              required: VALIDATIONS_FULLNAME.required,
-              minLength: { ...VALIDATIONS_FULLNAME.minLength },
-              maxLength: { ...VALIDATIONS_FULLNAME.maxLength },
-              pattern: { ...PATTERN_FULLNAME },
-            })}
-          />
-
-          <PhoneInput
-            className={styles.gridColumn}
-            type={SETTINGS_PHONE.type}
-            label={SETTINGS_PHONE.label}
-            placeholder={SETTINGS_PHONE.placeholder}
-            disabled={isActiveForm === false}
-            id="person_telephone"
-            autoComplete="off"
-            error={Boolean(errors.person_telephone?.message)}
-            errorMessage={errors.person_telephone?.message}
-            value={watch("person_telephone")}
-            {...register("person_telephone", {
-              required: "Данное поле обязательно",
-              minLength: {
-                message: "Минимальная длина номера 12 цифр",
-                value: 16,
-              },
-            })}
-          />
-
-          <Input
-            type={SETTINGS_EMAIL.type}
-            label={SETTINGS_EMAIL.label}
-            placeholder={SETTINGS_EMAIL.placeholder}
-            disabled
-            id="email"
-            autoComplete="on"
-            error={Boolean(errors.email?.message)}
-            errorMessage={errors.email?.message}
-            {...register("email", {
-              required: VALIDATIONS_EMAIL.required,
-              minLength: { ...VALIDATIONS_EMAIL.minLength },
-              maxLength: { ...VALIDATIONS_EMAIL.maxLength },
-              pattern: { ...PATTERN_EMAIL },
-            })}
-          />
-
-          <PasswordInput
-            className={styles.gridColumn}
-            type={SETTINGS_PREVIOUS_PASSWORD.type}
-            label={SETTINGS_PREVIOUS_PASSWORD.label}
-            placeholder={SETTINGS_PREVIOUS_PASSWORD.placeholder}
-            disabled={isActiveForm === false}
-            id="password"
-            autoComplete="off"
-            error={Boolean(errors.currentPass?.message)}
-            errorMessage={errors.currentPass?.message}
-            {...register("currentPass", {
-              validate: { ...VALIDATIONS_PASSWORD.validate },
-              required: {
-                value: Boolean(currentPass) || Boolean(newPass) || Boolean(newPassRepeat),
-                message: VALIDATIONS_PASSWORD.required,
-              },
-              minLength: { ...VALIDATIONS_PASSWORD.minLength },
-              maxLength: { ...VALIDATIONS_PASSWORD.maxLength },
-              pattern: { ...PATTERN_PASSWORD },
-            })}
-          />
-
-          <PasswordInput
-            type={SETTINGS_NEW_PASSWORD.type}
-            label={SETTINGS_NEW_PASSWORD.label}
-            placeholder={SETTINGS_NEW_PASSWORD.placeholder}
-            disabled={isActiveForm === false}
-            id="newPassword"
-            autoComplete="off"
-            error={Boolean(errors.newPass?.message)}
-            errorMessage={errors.newPass?.message}
-            {...register("newPass", {
-              required: {
-                value: Boolean(currentPass) || Boolean(newPass) || Boolean(newPassRepeat),
-                message: VALIDATIONS_PASSWORD.required,
-              },
-              minLength: { ...VALIDATIONS_PASSWORD.minLength },
-              maxLength: { ...VALIDATIONS_PASSWORD.maxLength },
-              pattern: { ...PATTERN_PASSWORD },
-              validate: {
-                differentPassword: (value = "") => {
-                  if (value.length && value === currentPass) {
-                    return "Новый пароль должен отличаться";
-                  }
-                },
-                ...VALIDATIONS_PASSWORD.validate,
-              },
-            })}
-          />
-
-          <PasswordInput
-            type={SETTINGS_NEW_PASSWORD_REPEAT.type}
-            label={SETTINGS_NEW_PASSWORD_REPEAT.label}
-            placeholder={SETTINGS_NEW_PASSWORD_REPEAT.placeholder}
-            disabled={isActiveForm === false}
-            id="newPasswordRepeat"
-            autoComplete="off"
-            error={Boolean(errors.newPassRepeat?.message)}
-            errorMessage={errors.newPassRepeat?.message}
-            {...register("newPassRepeat", {
-              required: {
-                value: Boolean(currentPass) || Boolean(newPass) || Boolean(newPassRepeat),
-                message: VALIDATIONS_PASSWORD.required,
-              },
-              minLength: { ...VALIDATIONS_PASSWORD.minLength },
-              maxLength: { ...VALIDATIONS_PASSWORD.maxLength },
-              pattern: { ...PATTERN_PASSWORD },
-              validate: {
-                passwordMatches: (value = "") => {
-                  if (value.length && value !== newPass) {
-                    return "Пароли не совпадают";
-                  }
-                },
-                ...VALIDATIONS_PASSWORD.validate,
-              },
-            })}
-          />
-        </div>
-
-        {buttonsFormComponent({
-          editedForm: isActiveForm,
-          isLoading: isLoadingUsers || isLoadingPassword,
-          formParameter: { changesForm: isDirty, isValidFields: isValid },
-        })}
-
-        <div className={styles.wrapperCheckbox}>
-          <h2 className="subtitle2">Получать уведомления</h2>
-          {checkboxFormComponent({
-            editedForm: isActiveForm,
-            noticeEmail,
-            onClick: handleChangeNotice,
-          })}
-        </div>
+        </FormProvider>
 
         {deletingProfileComponent()}
       </form>
