+++ src/redux/api/generated.ts
@@ -9,8 +9,6 @@
  * ---------------------------------------------------------------
  */
 
-import { ReactNode } from "react";
-
 export interface TokenObtainPair {
   /**
    * Email
@@ -249,8 +247,6 @@ export interface Message {
 }
 
 export interface Chat {
-  name: ReactNode;
-  orderId: number;
   /** ID (может отсутствовавть при создании чата) */
   id?: number;
   /** Заказчик */
@@ -310,28 +306,48 @@ export interface ContractorAgreement {
   created_date?: string;
 }
 
-export interface OrderOffer {
-  name: ReactNode;
+export interface OfferModel {
   /** Id */
-  id?: number;
-  user_account?: UserAccount;
-  /** Order id */
-  order_id?: string;
-  /**
-   * Цена офера
-   * @maxLength 300
-   */
-  offer_price?: string;
+  offer_id: number;
+  /** Цена офера */
+  offer_price: string;
   /**
    * Время выполнения офера
    * @maxLength 300
    */
-  offer_execution_time?: string;
+  offer_execution_time: string;
   /**
    * Описание офера
+   * @maxLength 1500
+   */
+  offer_description: string;
+  /** Статус */
+  status: "processed" | "viewed" | "selected" | "rejected";
+  /** Order id */
+  order_id: number;
+  /**
+   * Дата формирования заказа
    * @maxLength 300
    */
-  offer_description?: string;
+  order_time?: string;
+  /**
+   * Название заказа
+   * @maxLength 300
+   */
+  order_name?: string;
+  /**
+   * Описание заказа
+   * @maxLength 1500
+   */
+  order_description?: string;
+  /** Прикреплённые изображения */
+  order_images?: string[];
+  /** Цвет аватарки заказчика */
+  user_color: string;
+  /** ??? */
+  worksheet?: string | null;
+  /** ID чата */
+  chat_id: number;
 }
 
 export interface AllOrdersClient {
